- chg: updated build system to Gradle Kotlin DSL - chg: adaptation to recent API changes

+++ common/src/main/java/com/asamm/locus/addon/wear/common/communication/containers/trackrecording/TrackRecordingValue.java
@@ -199,7 +199,7 @@ public class TrackRecordingValue extends TimeStampStorable {
         /**
          * current hrm value
          */
-        private int hrm = 0;
+        private short hrm = 0;
         private float altitude = Float.NaN;
 
         public ExtendedTrackInfo() {
@@ -211,7 +211,7 @@ public class TrackRecordingValue extends TimeStampStorable {
             read(data);
         }
 
-        public ExtendedTrackInfo(Float speed, int hr, float altitude) {
+        public ExtendedTrackInfo(Float speed, short hr, float altitude) {
             this();
             mSpeed = speed == null ? INVALID_SPEED : speed;
             this.hrm = hr;
